separated main App.jsx into Header and Body components
