+++ client/src/components/Header.jsx
@@ -1,0 +1,12 @@
+import React, { useState } from 'react';
+
+const Header = ( { setDisplay } ) => {
+
+  return (
+    <header>
+      Some text
+    </header>
+  )
+}
+
+export default Header;
